Group 1 Ticketing System API/requirements.txt
sqlalchemy_utils
Faker
fastapi[all]
Flask